Register serializers in setup(), move code out of init(), rename Setup to AstralItemGroups
