+++ src/main/java/com/alan199921/astral/Astral.java
@@ -21,9 +21,13 @@ import com.alan199921.astral.commands.AstralCommands;
 import com.alan199921.astral.configs.AstralConfig;
 import com.alan199921.astral.entities.AstralEntityRegistry;
 import com.alan199921.astral.entities.PhysicalBodyEntityRenderer;
+import com.alan199921.astral.items.AstralItemGroups;
 import com.alan199921.astral.items.AstralItems;
 import com.alan199921.astral.network.AstralNetwork;
-import com.alan199921.astral.setup.AstralSetup;
+import com.alan199921.astral.worldgen.OverworldVegetation;
+import net.minecraft.item.ItemGroup;
+import net.minecraft.item.ItemStack;
+import net.minecraft.network.datasync.DataSerializers;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.client.event.ModelRegistryEvent;
 import net.minecraftforge.common.MinecraftForge;
@@ -42,13 +46,24 @@ import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import net.minecraftforge.fml.loading.FMLPaths;
 import net.minecraftforge.fml.network.simple.SimpleChannel;
 
+import static com.alan199921.astral.serializing.AstralSerializers.OPTIONAL_GAME_PROFILE;
+import static com.alan199921.astral.serializing.AstralSerializers.OPTIONAL_ITEMSTACK_HANDLER;
+
 // The value here should match an entry in the META-INF/mods.toml file
 @Mod(Astral.MOD_ID)
 public class Astral {
     public static final String MOD_ID = "astral";
     public static final SimpleChannel INSTANCE = AstralNetwork.getNetworkChannel();
+    //Astral ItemGroup using  a Snowberry Bush as an icon
+    public final ItemGroup astralItems = new ItemGroup("astral") {
+        @Override
+        public ItemStack createIcon() {
+            return new ItemStack(AstralBlocks.SNOWBERRY_BUSH.get());
+        }
+    };
 
-    public static final AstralSetup setup = new AstralSetup();
+
+    public static final AstralItemGroups setup = new AstralItemGroups();
 
     public Astral() {
         // Register the setup method for modloading
@@ -67,7 +82,13 @@ public class Astral {
     }
 
     private void setup(final FMLCommonSetupEvent event) {
-        setup.init();
+        //Initializes worldgen
+        OverworldVegetation.addOverworldVegetation();
+
+        //Register Serializers
+        DataSerializers.registerSerializer(OPTIONAL_GAME_PROFILE);
+        DataSerializers.registerSerializer(OPTIONAL_ITEMSTACK_HANDLER);
+
     }
 
     // You can use EventBusSubscriber to automatically subscribe events on the contained class (this is subscribing to the MOD
